==> code/1_indicators/1-text_length_descriptives.Rmd <==
---
title: "Text Length - Descriptives"
author: "Johanna Mehler"
date: "`r Sys.Date()`"
output: 
  html_document:
    toc: true               # Enable Table of Contents
    toc_float:              # Make the TOC floating
      collapsed: false      # Do not collapse TOC by default
      smooth_scroll: true   # Smooth scrolling to sections
    number_sections: true   # Numbered sections
    theme: readable         # Use a predefined theme for styling (optional)
    highlight: tango        # Syntax highlighting style (optional)
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = FALSE)
```

```{r}
library(tidyverse)
library(ggplot2)
library(here)

# load reprocessed data
data <- read_csv(here("data/batches.csv"))
```


## Text length in words 

```{r}
# Calculate Word Counts
data_wordcount <- data %>%
  mutate(word_count_hatedef = strsplit(as.character(hate_definition), " ") %>%
           sapply(length))

# Create Distribution Plot for open_hatedefinition

# Adjust binwidth for open_hatedefinition
binwidth_hatedef = max(1, floor((392 - 1) / 30))  # for a reasonable number of bins

# Plot for open_hatedefinition
ggplot(data_wordcount, aes(x = word_count_hatedef)) +
  geom_histogram(binwidth = binwidth_hatedef, fill = "skyblue", color = "black") +
  xlim(c(0, 400)) +  # Adjust x-axis limits
  labs(title = "Word Count Distribution: open_hatedefinition",
       x = "Word Count",
       y = "Frequency")
```

Almost all answers have a length below 100.


## Text length in characters

```{r}
data_text_length <- read_csv(here("data/text_length.csv"))

data_text_length %>% 
  ggplot(aes(x = text_length)) +
  geom_histogram(fill = "pink", color = "black") +
  labs(title = "Character Count Distribution: open_hatedefinition",
       x = "Character Count",
       y = "Frequency")

data_text_length %>% 
  ggplot(aes(x = text_length_log2)) +
  geom_histogram(fill = "pink", color = "black") +
  labs(title = "Character Count Distribution: open_hatedefinition (log2)",
       x = "Character Count",
       y = "Frequency")
```
