==> code/1_indicators/2-readability_descriptives.Rmd <==
---
title: "Readability Descriptives"
author: "Johanna Mehler"
date: "`r Sys.Date()`"
output: 
  html_document:
    toc: true               # Enable Table of Contents
    toc_float:              # Make the TOC floating
      collapsed: false      # Do not collapse TOC by default
      smooth_scroll: true   # Smooth scrolling to sections
    number_sections: true   # Numbered sections
    theme: readable         # Use a predefined theme for styling (optional)
    highlight: tango        # Syntax highlighting style (optional)
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = FALSE)
```

# Load data and packages

```{r}
library(tidyverse)
library(descr)
library(ggplot2)
library(here)

# pure readability scores. No analysis with oterh parameters
readability_data <- read_csv(here("data/readability_all_scores.csv"))

readability_scores <- readability_data %>% select(-ResponseId, -hate_definition)
```

# Check frequencies and distributions

```{r}
descr(readability_scores)
```

```{r}
readability_long <- pivot_longer(readability_scores, cols = everything(), names_to = "variable", values_to = "score")

# Now, we can plot the density of each variable with its own color
ggplot(readability_long, aes(x = score, fill = variable)) +
  geom_density(alpha = 0.7) + # Adjust alpha for transparency to see overlapping areas
  scale_fill_brewer(palette = "Set1") + # This sets different colors for each variable
  theme_minimal() + # A cleaner theme for the plot
  labs(title = "Density of Readability Scores", x = "Score", y = "Density") +
  theme(legend.title = element_blank()) # Hide the legend title if desired
```

```{r}
# as facet wrap
ggplot(readability_long, aes(x = score)) +
  geom_density(aes(fill = variable), alpha = 0.7) + # Add fill based on variable to differentiate them
  scale_fill_brewer(palette = "Set1") + # This sets different colors for each variable
  facet_wrap(~variable, scales = "free_y") + # This creates a separate plot for each variable
  theme_minimal() + # A cleaner theme for the plot
  labs(title = "Density of Readability Scores by Algorithm", x = "Score", y = "Density") +
  theme(legend.position = "none") # Hide the legend as it's redundant in this layout
```

## Lowest average scores

```{r}
# Sort the dataset by Average_Grade_Level in ascending order to get the 15 lowest scores
lowest_scores <- readability_data %>%
  arrange(Average_Grade_Level) %>%
  slice(1:15)

# Print the texts corresponding to the lowest scores
print(lowest_scores$hate_definition)
```

## Highest average scores

```{r}
# Sort the dataset by Average_Grade_Level in descending order to get the 15 highest scores
highest_scores <- readability_data %>%
  arrange(desc(Average_Grade_Level)) %>%
  slice(1:15)

# Print the texts corresponding to the highest scores
print(highest_scores$hate_definition)
```
