==> code/1_indicators/4-leftright_prediction_descriptives.Rmd <==
---
title: "Political Orientation Prediction Error Descriptives"
author: "Johanna Mehler"
date: "`r Sys.Date()`"
output: 
  html_document:
    toc: true               # Enable Table of Contents
    toc_float:              # Make the TOC floating
      collapsed: false      # Do not collapse TOC by default
      smooth_scroll: true   # Smooth scrolling to sections
    number_sections: true   # Numbered sections
    theme: readable         # Use a predefined theme for styling (optional)
    highlight: tango        # Syntax highlighting style (optional)
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = FALSE)
```

# Load data and packages

```{r}
library(tidyverse)
library(descr)
library(ggplot2)
library(here)
library(GGally)

# read in combined final dataset
data <- read_csv(here("data/data_combined.csv"))

data_all <- read_csv(here("data/data_clean.csv"))

pol_error <- data %>% select(leftright_pred_error)
```

# Check frequencies and distributions

## Distribution of the "real" variable leftright in my samples 

Distribution of my sample (n = 1079), Distribution of full sample (n > 19,000)


```{r}
# Distribution of training sample (n = 400)
freq(data$leftright)

# Distribution of test sample (n > 19,000)
freq(data_all$leftright)
```

## Descriptives of the new leftright_pred_error

```{r}
# Pol_error
descr(pol_error)
```



```{r}
# Density of leftright_pred_error
ggplot(pol_error, aes(x = leftright_pred_error)) +
  geom_density(fill = "skyblue", alpha = 0.6) + # Adjust alpha for transparency
  theme_minimal() +
  labs(title = "Density of Leftright Pred error",
       x = "Leftright Pred error",
       y = "Density") +
  theme(legend.title = element_blank()) # Hide the legend title
```

## Highest errors

```{r}
# Sort the dataset by Average_Grade_Level in ascending order to get the 15 lowest errors
highest_errors <- data %>%
  arrange(leftright_pred_error) %>%
  slice(1:15)

# Print Definition and leftright value
for (i in 1:nrow(highest_errors)) {
  cat("# Leftright Value:", highest_errors$leftright[i], "\n",
      "Open Hate Definition:", highest_errors$hate_definition[i], "\n\n")
}
```

## Lowest errors

```{r}
# Sort the dataset by Average_Grade_Level in descending order to get the 15 highest errors
lowest_errors <- data %>%
  arrange(desc(leftright_pred_error)) %>%
  slice(1:15)

# Print the texts corresponding to the highest errors
for (i in 1:nrow(lowest_errors)) {
  cat("# Leftright Value:", lowest_errors$leftright[i], "\n",
      "Open Hate Definition:", lowest_errors$hate_definition[i], "\n\n")
}
```


# Feature Importance

```{r}
library(knitr)
library(kableExtra)

coefficients <- read_csv(here("data/coefficients.csv"))

# Create table
styled_table <- kable(coefficients, format = "html", caption = "Coefficients of Predictor Variables", digits = 3) %>%
  kable_styling(bootstrap_options = c("striped", "hover", "condensed"), full_width = F, font_size = 12) %>%
  column_spec(1, bold = FALSE, color = "black")

# Print the table
styled_table

```


# Difference of Prediction and real Values 

```{r}
leftright_data <- data %>% select(leftright, leftright_pred_error, academic_status) %>% drop_na()

p <- ggpairs(leftright_data,  # Data frame
       # columns = indicators, # Columns
        aes(color = academic_status,  # Color by group
            alpha = 0.5),
        upper = list(continuous = wrap("cor", size = 2.7)
        ))

p
```
